added, but still problem with fields
